backend/requirements.txt
alembic==1.13.0
bcrypt==4.1.1
blinker==1.7.0
click==8.1.7
contourpy==1.2.0
cvxopt==1.3.2
cycler==0.12.1
deprecation==2.1.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.46.0
graphviz==0.20.1
install==1.3.5
intervaltree==3.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pm4py==2.7.8.4
pydotplus==2.0.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
scipy==1.11.4
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
StringDist==1.0.9
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==3.0.1
WTForms==3.1.1
